pyproject.toml
[project]
name = "websitegpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "requests",
    "beautifulsoup4",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "faiss-cpu==1.7.4",
    "numpy<2.0",
    "openai",
    "unstructured",
    "tiktoken",
    "streamlit",
    "pinecone",
    "python-dotenv>=1.2.1",
]
